# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Opt-Sim API",
    "description": "Prototype API for the Opt-Sim microstructure design workspace.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Run",
        "operationId": "read_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorPayload"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/draft": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Run Draft",
        "operationId": "read_run_draft_api_runs__run_id__draft_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDraft"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorPayload"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/targets": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read Run Targets",
        "operationId": "read_run_targets_api_runs__run_id__targets_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TargetAsset"
                  },
                  "title": "Response Read Run Targets Api Runs  Run Id  Targets Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorPayload"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/timeline": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Read Run Timeline",
        "operationId": "read_run_timeline_api_runs__run_id__timeline_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimelineEvent"
                  },
                  "title": "Response Read Run Timeline Api Runs  Run Id  Timeline Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorPayload"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/candidates": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Read Run Candidates",
        "operationId": "read_run_candidates_api_runs__run_id__candidates_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateSolution"
                  },
                  "title": "Response Read Run Candidates Api Runs  Run Id  Candidates Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorPayload"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/constraints": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Read Run Constraints",
        "operationId": "read_run_constraints_api_runs__run_id__constraints_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConstraintCheck"
                  },
                  "title": "Response Read Run Constraints Api Runs  Run Id  Constraints Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorPayload"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/export-estimate": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Read Run Export Estimate",
        "operationId": "read_run_export_estimate_api_runs__run_id__export_estimate_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportEstimate"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorPayload"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorPayload": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorPayload"
      },
...
      "CandidateMetric": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "CandidateMetric"
      },
      "CandidateParameter": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "CandidateParameter"
      },
      "CandidateSolution": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "selected": {
            "type": "boolean",
            "title": "Selected",
            "default": false
          },
          "status": {
            "type": "string",
            "enum": [
              "Recommended",
              "Robust",
              "Watch",
              "Blocked"
            ],
            "title": "Status"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/CandidateParameter"
            },
            "type": "array",
            "title": "Parameters"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/CandidateMetric"
            },
            "type": "array",
            "title": "Metrics"
          },
          "targetColorHex": {
            "type": "string",
            "title": "Targetcolorhex"
          },
          "simulatedColorHex": {
            "type": "string",
            "title": "Simulatedcolorhex"
          },
          "processPlusColorHex": {
            "type": "string",
            "title": "Processpluscolorhex"
          },
          "processMinusColorHex": {
            "type": "string",
            "title": "Processminuscolorhex"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rank",
          "group",
          "status",
          "parameters",
          "metrics",
          "targetColorHex",
          "simulatedColorHex",
          "processPlusColorHex",
          "processMinusColorHex",
          "rationale"
        ],
        "title": "CandidateSolution"
      },
      "ConstraintCheck": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "state": {
            "type": "string",
            "enum": [
              "pass",
              "warning",
              "fail",
              "unknown"
            ],
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "detail",
          "state"
        ],
        "title": "ConstraintCheck"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ExportEstimate": {
        "properties": {
          "dimensions": {
            "type": "string",
            "title": "Dimensions"
          },
          "fileSize": {
            "type": "string",
            "title": "Filesize"
          },
          "tilePlan": {
            "type": "string",
            "title": "Tileplan"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "tileProgress": {
            "type": "string",
            "title": "Tileprogress"
          }
        },
        "type": "object",
        "required": [
          "dimensions",
          "fileSize",
          "tilePlan",
          "format",
          "progress",
          "tileProgress"
        ],
        "title": "ExportEstimate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "Validating",
              "Simulating",
              "Ranking",
              "Needs approval",
              "Exporting",
              "Complete",
              "Failed"
            ],
            "title": "Status"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          },
          "warning": {
            "type": "boolean",
            "title": "Warning",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "updatedAt"
        ],
        "title": "RunSummary"
      },
      "TargetAsset": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "enum": [
              "color",
              "image",
              "multi-view"
            ],
            "title": "Type"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "swatchHex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Swatchhex"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "detail"
        ],
        "title": "TargetAsset"
      },
      "TimelineEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "agent",
              "tool",
              "warning",
              "approval",
              "result"
            ],
            "title": "Type"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "meta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Draft",
                  "Validating",
                  "Simulating",
                  "Ranking",
                  "Needs approval",
                  "Exporting",
                  "Complete",
                  "Failed"
                ]
              },
              {
                "type": "string",
                "const": "Running"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "actionLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actionlabel"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "label",
          "title",
          "body"
        ],
        "title": "TimelineEvent"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkspaceDraft": {
        "properties": {
          "requirementText": {
            "type": "string",
            "title": "Requirementtext"
          },
          "targetLabel": {
            "type": "string",
            "title": "Targetlabel"
          },
          "targetValue": {
            "type": "string",
            "title": "Targetvalue"
          },
          "heightWindow": {
            "type": "string",
            "title": "Heightwindow"
          },
          "exportMode": {
            "type": "string",
            "title": "Exportmode"
          }
        },
        "type": "object",
        "required": [
          "requirementText",
          "targetLabel",
          "targetValue",
          "heightWindow",
          "exportMode"
        ],
        "title": "WorkspaceDraft"
      },
...
    }
  }
}